🔨️ Refactored out project listing to its own controller
