+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/projects/metadata/ProjectLineListController.java
@@ -16,7 +16,6 @@ import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectMetadataTemp
 import ca.corefacility.bioinformatics.irida.model.project.Project;
 import ca.corefacility.bioinformatics.irida.model.sample.MetadataTemplate;
 import ca.corefacility.bioinformatics.irida.model.sample.MetadataTemplateField;
-import ca.corefacility.bioinformatics.irida.ria.web.projects.ProjectControllerUtils;
 import ca.corefacility.bioinformatics.irida.ria.web.projects.settings.ProjectBaseController;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
@@ -32,15 +31,13 @@ public class ProjectLineListController extends ProjectBaseController {
 
 	private final ProjectService projectService;
 	private final MetadataTemplateService metadataTemplateService;
-	private final ProjectControllerUtils projectControllerUtils;
 	private final MessageSource messageSource;
 
 	@Autowired
 	public ProjectLineListController(ProjectService projectService, MetadataTemplateService metadataTemplateService,
-			ProjectControllerUtils utils, MessageSource messageSource) {
+			MessageSource messageSource) {
 		this.projectService = projectService;
 		this.metadataTemplateService = metadataTemplateService;
-		this.projectControllerUtils = utils;
 		this.messageSource = messageSource;
 	}
 
@@ -79,7 +76,14 @@ public class ProjectLineListController extends ProjectBaseController {
 	@RequestMapping("/linelist-templates")
 	public String getLinelistTemplatePage(@PathVariable Long projectId, Model model, Locale locale) {
 		Project project = projectService.read(projectId);
-		model.addAttribute("templates", projectControllerUtils.getTemplateNames(locale, project));
+		List<ProjectMetadataTemplateJoin> metadataTemplatesForProject = metadataTemplateService.getMetadataTemplatesForProject(
+				project);
+		List<Map<String, String>> templates = new ArrayList<>();
+		for (ProjectMetadataTemplateJoin projectMetadataTemplateJoin : metadataTemplatesForProject) {
+			MetadataTemplate template = projectMetadataTemplateJoin.getObject();
+			templates.add(ImmutableMap.of("label", template.getLabel(), "id", String.valueOf(template.getId())));
+		}
+		model.addAttribute("templates", templates);
 		return "projects/project_linelist_template";
 	}
